style(runtime): keep agent spawn under line budget
